[issue-406] extract json parser and restructure test order
