+++ src/parser/json/json_parser.py
@@ -1,0 +1,22 @@
+#  Copyright (c) 2022 spdx contributors
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#      http://www.apache.org/licenses/LICENSE-2.0
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+import json
+from typing import Dict
+
+from src.model.document import Document
+from src.parser.jsonlikedict.json_like_dict_parser import JsonLikeDictParser
+
+
+def parse_from_file(file_name: str) -> Document:
+    with open(file_name) as file:
+        input_doc_as_dict: Dict = json.load(file)
+
+    return JsonLikeDictParser().parse(input_doc_as_dict)
