+++ src/parser/jsonlikedict/json_like_dict_parser.py
@@ -9,6 +9,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 import json
+from typing import Dict
 
 from src.model.document import Document
 from src.parser.error import SPDXParsingError
@@ -44,24 +45,21 @@ class JsonLikeDictParser:
         self.relationship_parser = RelationshipParser()
         self.annotation_parser = AnnotationParser()
 
-    def parse(self, filename: str) -> Document:
+    def parse(self, json_like_dict: Dict) -> Document:
 
-        with open(filename) as file:
-            input_doc_as_dict = json.load(file)
-
-        fields_to_parse = [("creation_info", input_doc_as_dict, self.creation_info_parser.parse_creation_info, False),
-                           ("packages", input_doc_as_dict.get("packages"), lambda x: parse_list_of_elements(x,
-                                                                                                            self.package_parser.parse_package,
-                                                                                                            self.package_parser.logger), True),
-                           ("files", input_doc_as_dict.get("files"), lambda x: parse_list_of_elements(x,
-                                                                                                      self.file_parser.parse_file,
-                                                                                                      self.file_parser.logger), True),
-                           ("annotations", input_doc_as_dict, self.annotation_parser.parse_all_annotations, True),
-                           ("snippets", input_doc_as_dict.get("snippets"), lambda x: parse_list_of_elements(x,
-                                                                                                            self.snippet_parser.parse_snippet,
-                                                                                                            self.snippet_parser.logger), True),
-                           ("relationships", input_doc_as_dict, self.relationship_parser.parse_all_relationships, True),
-                           ("extracted_licensing_info", input_doc_as_dict.get("hasExtractedLicensingInfos"),
+        fields_to_parse = [("creation_info", json_like_dict, self.creation_info_parser.parse_creation_info, False),
+                           ("packages", json_like_dict.get("packages"), lambda x: parse_list_of_elements(x,
+                                                                                                         self.package_parser.parse_package,
+                                                                                                         self.package_parser.logger), True),
+                           ("files", json_like_dict.get("files"), lambda x: parse_list_of_elements(x,
+                                                                                                   self.file_parser.parse_file,
+                                                                                                   self.file_parser.logger), True),
+                           ("annotations", json_like_dict, self.annotation_parser.parse_all_annotations, True),
+                           ("snippets", json_like_dict.get("snippets"), lambda x: parse_list_of_elements(x,
+                                                                                                         self.snippet_parser.parse_snippet,
+                                                                                                         self.snippet_parser.logger), True),
+                           ("relationships", json_like_dict, self.relationship_parser.parse_all_relationships, True),
+                           ("extracted_licensing_info", json_like_dict.get("hasExtractedLicensingInfos"),
                             lambda x: parse_list_of_elements(x,
                                                              self.extracted_licensing_info_parser.parse_extracted_licensing_info,
                                                              self.extracted_licensing_info_parser.logger), True)]
